libzmqutil: rename reactor.c to zwatcher.c
Problem: the zmq watcher is implemented in an ambigously named
source file.

Use zwatcher.[ch] instead of reactor.[ch]
Rename unit test too.
Update include directives.
